requirements.txt
ak-gpapi==1.0.2
rich==13.7.0
typer==0.9.0
ndg-httpsclient==0.5.1
pyasn1==0.4.8
certifi==2020.6.20
urllib3==1.25.10
